+++ pkgs/native_assets_builder/test/build_runner/build_runner_reusability_test.dart
@@ -2,6 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:file/local.dart';
 import 'package:native_assets_builder/src/build_runner/build_runner.dart';
 import 'package:test/test.dart';
 
@@ -25,12 +26,18 @@ void main() async {
       final buildRunner = NativeAssetsBuildRunner(
         logger: logger,
         dartExecutable: dartExecutable,
+        fileSystem: const LocalFileSystem(),
       );
 
+      final targetOS = OS.current;
+      const defaultMacOSVersion = 13;
       BuildConfigBuilder configCreator() => BuildConfigBuilder()
         ..setupCodeConfig(
           targetArchitecture: Architecture.current,
           targetOS: OS.current,
+          macOSConfig: targetOS == OS.macOS
+              ? MacOSConfig(targetVersion: defaultMacOSVersion)
+              : null,
           linkModePreference: LinkModePreference.dynamic,
         );