+++ pkgs/native_assets_builder/CHANGELOG.md
@@ -1,9 +1,20 @@
-## 0.10.0-wip
+## 0.10.2
+
+- Export types (fix for 0.10.1).
+
+## 0.10.1
+
+- Pass in the environment for hook invocations.
+
+## 0.10.0
 
 - Removed support for dry run (Flutter no long requires it).
 - Various fixes to caching.
 - **Breaking change** `BuildConfig.targetOS` is now only provided if
   `buildAssetTypes` contains the code asset.
+- **Breaking change** `NativeAssetsBuildRunner` and `PackageLayout` now take a
+  `FileSystem` from `package:file/file.dart`s.
+- Bump `package:native_assets_cli` to 0.10.0.
 
 ## 0.9.0